fix(runtime): match map import success result
Return 1 after a successful SQL_IMPORT_MAP_XML operation to match the snake reference implementation. Cover normal, empty, and maximum-row successful imports with regression assertions.

+++ crates/era-runtime/src/session/interaction/sql/response.rs
@@ -462,7 +462,7 @@ impl RuntimeSession {
                 era_runtime_protocol::SqlResultV1::MapImported { .. },
             ) => {
                 self.sql.release_connection(&connection_key);
-                self.finish_sql_value(request, VmValue::Integer(0))
+                self.finish_sql_value(request, VmValue::Integer(1))
             }
             (
                 SqlServiceContinuation::Disconnect {
